Some Updates
- Updated RequestHeader.java
- Reworked whitelist and blacklist system (not finally in this commit)
- Updated VerificationListener
- Implemented bots.discord.pw
- Improoved SafeMessage.java

+++ src/main/java/fun/rubicon/listener/VerificationListener.java
@@ -29,6 +29,7 @@ public class VerificationListener extends ListenerAdapter {
 
     @Override
     public void onGuildMessageReceived(GuildMessageReceivedEvent event) {
+        if(event.getAuthor().isBot()) return;
         if (!CommandVerification.setups.containsKey(event.getGuild())) return;
         if (!CommandVerification.setups.get(event.getGuild()).author.equals(event.getAuthor())) return;
         CommandVerification.VerificationSetup setup = CommandVerification.setups.get(event.getGuild());
@@ -63,6 +64,7 @@ public class VerificationListener extends ListenerAdapter {
 
     @Override
     public void onGuildMemberJoin(GuildMemberJoinEvent event) {
+        if(event.getUser().isBot()) return;
         if (!RubiconBot.getMySQL().verificationEnabled(event.getGuild())) return;
         TextChannel channel = event.getGuild().getTextChannelById(RubiconBot.getMySQL().getVerificationValue(event.getGuild(), "channelid"));
         Message message = SafeMessage.sendMessageBlocking(channel, RubiconBot.getMySQL().getVerificationValue(event.getGuild(), "text").replace("%user%", event.getUser().getAsMention()).replace("%guild%", event.getGuild().getName()));
@@ -81,6 +83,7 @@ public class VerificationListener extends ListenerAdapter {
 
     @Override
     public void onGuildMemberLeave(GuildMemberLeaveEvent event) {
+        if(event.getUser().isBot()) return;
         if (VerificationKickHandler.VerifyKick.exists(event.getMember())) {
             VerificationKickHandler.VerifyKick kick = VerificationKickHandler.VerifyKick.fromMember(event.getMember(), true);
             event.getJDA().getTextChannelById(RubiconBot.getMySQL().getVerificationValue(event.getGuild(), "channelid")).getMessageById(kick.getMessageId()).complete().delete().queue();
